jobs/quicksuggest-2-bq/requirements.txt
click==7.1.2
google-cloud-bigquery==2.30.1
kinto-http==10.7.0
pytest==6.0.2
pytest-black==0.3.11
pytest-flake8==1.0.6
pytest-mock==3.6.1
